integrate farms in new sidebar
