+++ packages/playground/src/weblets/full_vm.vue
@@ -165,7 +165,7 @@ import SelectFarmManager from "../components/select_farm_manager.vue";
 import { useLayout } from "../components/weblet_layout.vue";
 import { useProfileManager } from "../stores";
 import type { solutionFlavor as SolutionFlavor } from "../types";
-import { type Farm, type Flist, ProjectName } from "../types";
+import { type FarmInterface, type Flist, ProjectName } from "../types";
 import { deployVM, type Disk } from "../utils/deploy_vm";
 import { getGrid } from "../utils/grid";
 import { normalizeError } from "../utils/helpers";
@@ -207,7 +207,7 @@ const planetary = ref(true);
 const wireguard = ref(false);
 const dedicated = ref(false);
 const certified = ref(false);
-const farm = ref() as Ref<Farm>;
+const farm = ref() as Ref<FarmInterface>;
 const loadingFarm = ref(false);
 const disks = ref<Disk[]>([]);
 const network = ref();
