+++ packages/playground/src/weblets/tf_presearch.vue
@@ -130,7 +130,7 @@ import { type Ref, ref } from "vue";
 import Network from "../components/networks.vue";
 import { useLayout } from "../components/weblet_layout.vue";
 import { useProfileManager } from "../stores";
-import { type Farm, type Flist, ProjectName } from "../types";
+import { type FarmInterface, type Flist, ProjectName } from "../types";
 import { deployVM } from "../utils/deploy_vm";
 import { getGrid } from "../utils/grid";
 import { normalizeError } from "../utils/helpers";
@@ -150,7 +150,7 @@ const memory = 512;
 const rootFilesystemSize = rootFs(cpu, memory);
 const loadingFarm = ref(false);
 const dockerDiskSize = 10;
-const farm = ref() as Ref<Farm>;
+const farm = ref() as Ref<FarmInterface>;
 const farmName = ref();
 const privateRestoreKey = ref("");
 const publicRestoreKey = ref("");
